requirements.txt
pydantic==2.1.1
pytest==7.4.0
python-dotenv==1.0.0
requests==2.31.0
schedule==1.2.0
slack-sdk==3.21.3
markdownify~=0.11.6
urllib3~=2.0.4
pydantic_settings